remove tabs
