+++ frontend/src/components/DefaultLayout.vue
@@ -87,7 +87,7 @@ export default class DefaultLayout extends Vue {
   width: 100%;
   height: 100%;
   /* Um auf der Y-Achse direkt unter der App Bar zu liegen */
-  padding-top: 92px;
+  padding-top: 50px;
   position: fixed;
   /* Um auf der Z-Achse direkt unter der App Bar zu liegen */
   z-index: 4;
