+++ frontend/src/views/Bauvorhaben.vue
@@ -256,7 +256,7 @@ export default class Bauvorhaben extends Mixins(
       Toaster.toast(message, level);
     }
 
-    this.$router.push({ name: "karte" });
+    this.$router.push({ name: "home" });
   }
 
   /**
