--- frontend/src/views/InfrastruktureinrichtungenUebersicht.vue
@@ -1,169 +1,0 @@
-<template>
-  <default-layout>
-    <template #content>
-      <div
-        v-if="infrastruktureinrichtungen.length > 0"
-        class="py-12"
-      >
-        <!-- eslint-disable vue/no-unused-vars -->
-        <v-hover
-          v-for="(item, index) in infrastruktureinrichtungen"
-          :key="item.id"
-          v-slot="{ hover }"
-        >
-          <v-card
-            outlined
-            class="mb-4 transition-swing"
-            :elevation="hover ? 4 : 0"
-            @click="editInfrastruktureinrichtung(item)"
-          >
-            <v-card-title :id="'infrastruktureinrichtung_uebersicht_item_' + index + '_nameEinrichtung'">
-              {{ item.nameEinrichtung }}
-              <v-spacer />
-            </v-card-title>
-            <v-card-text>
-              <span :id="'infrastruktureinrichtung_uebersicht_item_' + index + '_infrastruktureinrichtungtyp'">
-                {{ getLookupValue(item.infrastruktureinrichtungTyp, infrastruktureinrichtungTypList) }}</span
-              >
-            </v-card-text>
-          </v-card>
-        </v-hover>
-      </div>
-      <loading
-        v-else
-        id="infrastruktureinrichtung_uebersicht_loading"
-        :success="backendAccessSuccessful"
-        name="Infrastruktureinrichtungen"
-      />
-    </template>
-    <template #action>
-      <v-spacer />
-      <div class="align-self-end">
-        <v-tooltip left>
-          <template #activator="{ on }">
-            <v-btn
-              id="infrastruktureinrichtung_uebersicht_infrastruktureinrichtung_erstellen_button"
-              slot="activator"
-              v-model="options"
-              dark
-              fab
-              x-large
-              color="secondary"
-              :disabled="!isEditable"
-              v-on="on"
-              @click="newInfrastruktureinrichtung"
-            >
-              <v-icon> mdi-plus </v-icon>
-            </v-btn>
-          </template>
-          <span v-if="options">Abbrechen</span>
-          <span v-else>Infrastruktureinrichtung erstellen</span>
-        </v-tooltip>
-      </div>
-    </template>
-  </default-layout>
-</template>
-
-<script lang="ts">
-import { Component, Mixins, Watch } from "vue-property-decorator";
-import router from "@/router";
-import {
-  InfrastruktureinrichtungListElementDto,
-  InfrastruktureinrichtungListElementsDto,
-  LookupEntryDto,
-} from "@/api/api-client/isi-backend";
-import InfrastruktureinrichtungApiRequestMixin from "@/mixins/requests/InfrastruktureinrichtungApiRequestMixin";
-import DefaultLayout from "@/components/DefaultLayout.vue";
-import _ from "lodash";
-import SecurityMixin from "@/mixins/security/SecurityMixin";
-
-@Component({
-  components: { DefaultLayout },
-})
-export default class InfrastruktureinrichtungenUebersicht extends Mixins(
-  InfrastruktureinrichtungApiRequestMixin,
-  SecurityMixin
-) {
-  private infrastruktureinrichtungen: Array<InfrastruktureinrichtungListElementDto> = [];
-
-  private options = false;
-
-  private backendAccessSuccessful: boolean | null = null;
-
-  get isEditable(): boolean {
-    return this.isRoleAdminOrSachbearbeitung();
-  }
-
-  get infrastruktureinrichtungTypList(): LookupEntryDto[] {
-    return this.$store.getters["lookup/infrastruktureinrichtungTyp"];
-  }
-
-  private getLookupValue(key: string, list: Array<LookupEntryDto>): string | undefined {
-    return !_.isUndefined(list) ? list.find((lookupEntry: LookupEntryDto) => lookupEntry.key === key)?.value : "";
-  }
-
-  @Watch("$store.state.search.resultInfrastruktureinrichtung", { immediate: true, deep: true })
-  private resultInfrastruktureinrichtungenChanged() {
-    const infrastruktureinrichtungenStore = this.$store.getters["search/resultInfrastruktureinrichtung"];
-    if (!_.isNil(infrastruktureinrichtungenStore)) {
-      this.infrastruktureinrichtungen = infrastruktureinrichtungenStore;
-    }
-  }
-
-  mounted(): void {
-    this.getInfrastruktureinrichtungen();
-  }
-
-  /**
-   * Liest alle Infrastruktureinrichtungen aus dem Backend
-   */
-  private async getInfrastruktureinrichtungen(): Promise<void> {
-    this.backendAccessSuccessful = null;
-    if (!this.$store.getters["search/isInitializedInfrastruktureinrichtung"]) {
-      await this.getInfrastruktureinrichtungenListElements(false)
-        .then((infrastruktureinrichtungListElementsDto: InfrastruktureinrichtungListElementsDto) => {
-          if (!_.isUndefined(infrastruktureinrichtungListElementsDto.listElements)) {
-            this.$store.dispatch(
-              "search/resultInfrastruktureinrichtung",
-              infrastruktureinrichtungListElementsDto.listElements
-            );
-            this.backendAccessSuccessful = true;
-          }
-        })
-        .catch(() => {
-          this.backendAccessSuccessful = false;
-        });
-    } else {
-      this.infrastruktureinrichtungen = this.$store.getters["search/resultInfrastruktureinrichtung"];
-      this.backendAccessSuccessful = true;
-    }
-  }
-
-  /**
-   * Routing zur View der existenten Infrastruktureinrichtung
-   *
-   * @param infrastruktureinrichtungListElementDto zum ermitteln der Route.
-   */
-  private editInfrastruktureinrichtung(
-    infrastruktureinrichtungListElementDto: InfrastruktureinrichtungListElementDto
-  ): void {
-    if (!_.isNil(infrastruktureinrichtungListElementDto.id)) {
-      router.push({
-        name: "editInfrastruktureinrichtung",
-        params: { id: infrastruktureinrichtungListElementDto.id },
-      });
-    }
-  }
-
-  /**
-   * Erstellung einer neuen Infrastruktureinrichtung.
-   */
-  private newInfrastruktureinrichtung(): void {
-    router.push({
-      name: "createInfrastruktureinrichtung",
-    });
-  }
-}
-</script>
-
-<style></style>
